update _numa_fit_instance_cell for the support of virtual_pmems
_numa_fit_instance_cell is used to get the instance cell based on
a fitted host cell. It will be called by numa_topology_filter to
find out fitted hosts for the instance, so we can update this to
support NUMA affinity. Also it will be called by instance_claim
when building an instance to update the instance's numa_topology.

Partiallly-Implements: blueprint virtual-persistent-memory

Co-Authored-By: He Jie Xu <hejie.xu@intel.com>

Change-Id: Ie6a3d8a903aa3c705b273998412a246c7c1e87fd